requirements.txt
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.20
plotly==5.17.0
plotly
numpy==1.24.3
